test : test 독립성 보장

+++ src/main/java/com/ll/exam/sbb/answer/AnswerRepository.java
@@ -9,7 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-public interface AnswerRepository extends JpaRepository<Answer, Integer>, RepositoryUtil {
+public interface AnswerRepository extends JpaRepository<Answer, Long>, RepositoryUtil {
 
     @Transactional
     @Modifying
